Move project storage code from data_connectors to storage module (#1397)
* WIP: starting move

* Moved more code

* Change path to /storage

* Fix alembic migration sequence

* register storage orms

* Fix type error

* Change api tags to storage

* Fix path in api.spec.yaml

* Fix validation errors

* Add `additionalProperties: false` to explicitly disallow extra param

* Fix query to select projects

* style fix

* Remove left-over defs in data_connector api.spec

- adopt test to new (pydantic) validation error

* Fix exclusiveMinimum spec

* Add missing types in api spec

they worked before because of being merged in from data_connectors
