+++ bases/renku_data_services/data_api/dependencies.py
@@ -40,7 +40,6 @@ from renku_data_services.data_connectors.db import (
 )
 from renku_data_services.data_connectors.deposits.envidat import EnvidatClient
 from renku_data_services.data_connectors.deposits.zenodo import ZenodoAPIClient
-from renku_data_services.data_connectors.project_storage_k8s import ProjectStorageK8s
 from renku_data_services.git.gitlab import DummyGitlabAPI, EmptyGitlabAPI, GitlabAPI
 from renku_data_services.k8s.client_interfaces import K8sClient
 from renku_data_services.k8s.clients import (
@@ -83,6 +82,8 @@ from renku_data_services.secrets.db import LowLevelUserSecretsRepo, UserSecretsR
 from renku_data_services.session.constants import BUILD_RUN_GVK, TASK_RUN_GVK
 from renku_data_services.session.db import SessionRepository
 from renku_data_services.session.k8s_client import ShipwrightClient
+from renku_data_services.storage.db import ProjectStorageRepository
+from renku_data_services.storage.project_storage_k8s import ProjectStorageK8s
 from renku_data_services.users.db import UserPreferencesRepository
 from renku_data_services.users.db import UserRepo as KcUserRepo
 from renku_data_services.users.dummy_kc_api import DummyKeycloakAPI
@@ -183,6 +184,7 @@ class DependencyManager:
     internal_token_mint: RenkuSelfTokenMint
     internal_scope_verifier: ScopeVerifier
     project_storage_k8s: ProjectStorageK8s
+    project_storage_repo: ProjectStorageRepository
 
     spec: dict[str, Any] = field(init=False, repr=False, default_factory=dict)
     app_name: str = "renku_data_services"
@@ -374,6 +376,13 @@ class DependencyManager:
             member_repo=member_repo,
         )
         project_storage_k8s = ProjectStorageK8s(config.nb_config.k8s_v2_client)
+        project_storage_repo = ProjectStorageRepository(
+            session_maker=config.db.async_session_maker,
+            authz=authz,
+            project_repo=project_repo,
+            group_repo=group_repo,
+            project_storage_config=config.project_storage_config,
+        )
         reprovisioning_repo = ReprovisioningRepository(session_maker=config.db.async_session_maker)
 
         git_repositories_repo = GitRepositoriesRepository(
@@ -430,7 +439,6 @@ class DependencyManager:
             project_repo=project_repo,
             group_repo=group_repo,
             search_updates_repo=search_updates_repo,
-            project_storage_config=config.project_storage_config,
         )
         data_connector_secret_repo = DataConnectorSecretRepository(
             session_maker=config.db.async_session_maker,
@@ -533,4 +541,5 @@ class DependencyManager:
             internal_token_mint=internal_token_mint,
             internal_scope_verifier=internal_scope_verifier,
             project_storage_k8s=project_storage_k8s,
+            project_storage_repo=project_storage_repo,
         )
