+++ components/renku_data_services/data_connectors/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2026-07-30T10:50:42+00:00
+#   timestamp: 2026-07-30T15:37:24+00:00
 
 from __future__ import annotations
 
@@ -561,10 +561,6 @@ class DataConnectorsGetQuery(PaginationRequest):
 
 class DataConnectorsGetParametersQuery(BaseAPISpec):
     params: DataConnectorsGetQuery | None = None
-
-
-class DataConnectorsStorageAllowGetParametersQuery(BaseAPISpec):
-    params: ProjectStorageAllowListQuery | None = None
 
 
 class DataConnectorLinksGetParametersQuery(BaseAPISpec):
